requirements.txt
PyQt6>=6.4.0
pyaudio>=0.2.13
numpy>=1.21.0
openai>=1.3.0
elevenlabs>=0.2.24
faster-whisper>=0.9.0
keyboard>=0.13.5
torch>=2.0.0
wave>=0.0.2
python-dotenv>=1.0.0
